fix: compile photo wallpaper animation math

+++ app/src/main/java/pl/szczodrzynski/edziennik/ui/aximo/AximoPhotoWallpaperDrawable.kt
@@ -59,12 +59,12 @@ class AximoPhotoWallpaperDrawable(
         val bw = b.width.toFloat()
         val bh = b.height.toFloat()
         val baseScale = maxOf(w / bw, h / bh)
-        val zoom = 1.045f + 0.025f * ((sin(p) + 1f) / 2f)
+        val zoom = 1.045f + 0.025f * ((sin(p.toDouble()).toFloat() + 1f) / 2f)
         val scale = baseScale * zoom
         val drawW = bw * scale
         val drawH = bh * scale
-        val driftX = (drawW - w) * (0.5f + 0.16f * sin(p * 0.7))
-        val driftY = (drawH - h) * (0.5f + 0.10f * sin(p * 0.55 + 1.2))
+        val driftX = (drawW - w) * (0.5f + 0.16f * sin((p * 0.7f).toDouble()).toFloat())
+        val driftY = (drawH - h) * (0.5f + 0.10f * sin((p * 0.55f + 1.2f).toDouble()).toFloat())
         val src = Rect(0, 0, b.width, b.height)
         val dst = RectF(-driftX, -driftY, -driftX + drawW, -driftY + drawH)
